refactort: create productsList component
